testing

+++ testing.py
@@ -362,7 +362,7 @@ for l in range(layers):
 
 #predict results
 model = load_model('../models_3/model_ref_back.h5')
-img = Image.fromarray(pred_back.astype(np.uint8))
+img = Image.fromarray(pred_back)
 pixels = img_to_array(pixels)
 pixels = (pixels - 127.5) / 127.5
 img = expand_dims(pixels, 0)
